split mapping into multiple files
